Moved Beans to seperate class
