+++ src/main/java/interview/project/Github/GithubProxyApplication.java
@@ -13,21 +13,4 @@ public class GithubProxyApplication {
     public static void main(String[] args) {
         SpringApplication.run(GithubProxyApplication.class, args);
     }
-
-    @Bean
-    public RestTemplate getRestTemplate() {
-        return new RestTemplate();
-    }
-
-    @Bean
-    public HttpEntity<String> getAuthorizationTokenHttpEntity(@Value("${use_personal_access_key}") Boolean usePersonalAccessKey,
-                                                              @Value("${github.personal.access.key}") String personalAccessKey) {
-        HttpHeaders headers = new HttpHeaders();
-
-        if (usePersonalAccessKey) {
-            headers.set("Authorization", "Bearer " + personalAccessKey);
-        }
-
-        return new HttpEntity<>("", headers);
-    }
 }
